% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cnv.plot.R
\name{cnv.plot}
\alias{cnv.plot}
\title{Create a copy number and structural variant plot.}
\usage{
cnv.plot(chr, snps, segments, svs)
}
\arguments{
\item{chr}{The chromosome to process}

\item{snps}{A \code{data.frame} containing the following columns Chr LRR BAF
Position}

\item{segments}{A \code{data.frame} containing the following columns Chr
Copy LRR}

\item{svs}{A \code{data.frame} containing the following columns Chr1 Chr2
Pos1 Pos2 Type Support}
}
\value{
none
}
\description{
Creates a figure containing structural variants (large INDELs,
translocations), B allele frequency, copy number changes and log R
ratios.  Each of these data types is supplied as a \code{data.frame} so
the function is agnostic about the source of the data.
}
\details{
While the function is agnostic about where the data comes from, the
\pkg{GenomicVis} package provides functions and examples to obtain data
from Illumina SNP-Chips that have been processed with GenomeStudio and
GAP and structural variants from BreakDancer.  Ultimately, however, this
data may come from any source (or even multiple sources) so long as it is
presented to the function via the three standadised \code{data.frame}s.

\describe{
\item{foo}{a foo bar}
}
}
\examples{
data(SNPExample)
data(CNVExample)
data(SVExample)
cnv.plot('18', SNPExample, CNVExample, SVExample)
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}

