% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenomicVis-package.R
\docType{data}
\name{CNVData}
\alias{CNVData}
\title{An example of a CNV data frame used by \code{cnv.heatmap}.}
\description{
An example of a CNV data frame used by \code{cnv.heatmap}.
}
\keyword{datasets}

