% Generated by roxygen2 (4.0.1): do not edit by hand
\name{default.cnv.heatmap.colour.map}
\alias{default.cnv.heatmap.colour.map}
\title{CNV heatmap colour scheme.}
\usage{
default.cnv.heatmap.colour.map()
}
\description{
A function that returns a colour scheme for the \code{cnv.heatmap}
function.  You need to provide a \code{data.frame} to \code{cnv.heatmap},
this function returns a \code{data.frame} that is suitable for the
example data contained in \code{CNVData}.
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}

