% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.breakdancer.R
\name{filter.breakdancer}
\alias{filter.breakdancer}
\title{Illustrative funtion to filter Breakdancer calls.}
\usage{
filter.breakdancer(sv, minTumour = 10, maxNormal = 4, minRatio = 5,
  chr = NA, conf = 95)
}
\arguments{
\item{sv}{A \code{data.frame} typically the returned value from
\code{read.breakdancer}}

\item{minTumour}{minimum number of reads supporting SV in the tumour
sample.}

\item{maxNormal}{maximum number of reads supporting SV in the normal
sample.}

\item{minRatio}{minimum ration between supporting reads in the tumour /
normal.}

\item{chr}{Chromosome to process}

\item{conf}{Minimum confidence (score) required.}
}
\value{
The filtered \code{data.frame}
}
\description{
This function is an example of filtering Breakdancer calls that have been
read with \code{read.breakdancer}.  It is illustrative only, and not
intended to necessarily represent the best way to filter Breakdancer
calls.
}
\author{
Lutz Krause <lutz.krause@qimrberghofer.edu.au>
}

