% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gap2cnv}
\alias{gap2cnv}
\title{Illustrative function for processing GAP data.}
\usage{
gap2cnv(gap)
}
\arguments{
\item{gap}{A \code{data.frame}, typically obtained by calling
\code{read.gap}.}
}
\value{
A \code{data.frame} suitable as input to the \code{cnv.heatmap}
function.
}
\description{
This function is an example of reading data output by GAP, and using it
as input for the \code{cnv.heatmap} function.
}
\examples{
\dontrun{
cnv <- gap2cnv(gap)
}
}
\author{
Lutz Krause <lutz.krause@qimrberghofer.edu.au>
}

