% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genes.in.region.R
\name{genes.in.region}
\alias{genes.in.region}
\title{Find gene in a genomic interval}
\usage{
genes.in.region(chr, begin, end, genes.gr)
}
\arguments{
\item{chr}{chr}

\item{begin}{begin}

\item{end}{end}

\item{genes.gr}{genes.gr}
}
\value{
A character containing a list of gene symbols (note it's not a
character vector, well it's a character vector of length 1).
}
\description{
This is an internal helper function for \code{cnv.heatmap}.
}
\author{
Lutz Krause <lutz.krause@qimrberghofer.edu.au>
}

