% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenomicVis-package.R
\docType{package}
\name{GenomicVis}
\alias{GenomicVis}
\alias{GenomicVis-package}
\title{Genomic Visualisation}
\description{
The \pkg{GenomicVis} package contains a collection of function for the
visualisation of genomic data typically derived from high-throughput
whole genome sequencing and SNP-Chip experiments.  Most of the functions
provided by this package are agnostic about where the data actually comes
from, that is, they required input in simple \code{data.frame} instances
that are not tied to a particular file format; however, we also provide
many examples of creating these simply \code{data.frame}s from standard
file formats such as VCFs and from common software tools such as GAP and
Breakdancer.
}
\details{
For further details, and examples, users are directed to the package
vignette: run \code{vignette('GenomicVis')} from the R prompt.
}

