% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kataegis.colours.R
\name{kataegis.colours}
\alias{kataegis.colours}
\title{Default colours for kataegis plot.}
\usage{
kataegis.colours()
}
\description{
Returns a named vector providing the default colours to use in kataegis
plots.
}

