% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kataegis}
\alias{kataegis}
\title{Calculate regions of kataegis from a VCF file.}
\usage{
kataegis(x, min.snvs = 6, max.mean.distance = 1000, pcf = FALSE,
  maxseg = NA, maxk = NA, ncpus = getOption("cl.cores", 1))
}
\arguments{
\item{x}{A \code{GRanges} instance containing SNVs.}

\item{min.snvs}{the minimum number of SNVs in a kataegis region}

\item{max.mean.distance}{the minimum mean genetic distance in a kataegis
region.}

\item{pcf}{Use piece-wise constant curve fitting algorithm to identify
kataegis. See `Details` below.}

\item{maxseg}{maximum segments in PCF algorithm, see Details.}

\item{maxk}{maximum length of segments in PCF algorithm, see Details.}

\item{ncpus}{Run the function in parallel using this number of CPUs.}
}
\value{
A \code{GRanges} object representing the regions of kataegis
identified
}
\description{
Identify regions of kataegis.  The input \code{GRanges} instance should
only contain SNV (i.e., no INDELs) that have a single ALT allele.  The
input \code{GRanges} instance must also contain a valid \code{Seqinfo}
class that gives the lengths of the chromosomes.  The easiest way to
obtain this from a VCF file is via the \code{read.vcf} function from this
package.
}
\details{
Details ...
}
\examples{
\dontrun{
vcf.file <- system.file('extdata', 'example.vcf',
  package = 'GenomicVis')
vcf <- read.vcf(vcf.file, 'GRCh37')
x <- rowData(vcf)
kat <- kataegis(x)
}
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}
\keyword{methods}

