% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.vcffiles}
\alias{list.vcffiles}
\title{List VCF files in a directory.}
\usage{
list.vcffiles(path = ".", full.names = FALSE, recursive = FALSE)
}
\arguments{
\item{path}{A character vector of full path names; the default corresponds
to the working directory, \code{getwd()}.  Tilde expansion (see
\code{path.expand}) is performed.}

\item{full.names}{a logical value. If \code{TRUE}, the directory path is
prepended to the file names to give a relative file path.  If \code{FALSE},
the file names (rather than paths) are returned.}

\item{recursive}{logical.  Should the listing recurse into directories?}
}
\value{
A character vector containing the names of the VCF files in the
specified directories, or \code{character(0)} if there were no VCF files.
}
\description{
This function produces a character vector of the names (or paths) of VCF
files in the named directory.
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}

