% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.kataegisList}
\alias{plot.kataegisList}
\title{Create a kataegis "rainfall" plot.}
\usage{
\method{plot}{kataegisList}(x, chr = NA, start = NA, end = NA,
  all = FALSE, alpha = 0.5, legend = TRUE, colours = kataegis.colours,
  ...)
}
\arguments{
\item{x}{a \code{kataegisList} object, usually created with the
\code{kataegis} function.}

\item{chr}{Optionally, only process this chromosome}

\item{start}{Optionally, only process from this start point (must be specify
together with \code{chr} and \code{end}).}

\item{end}{Optionally, only process until this point (must be specify
together with \code{chr} and \code{start}).}

\item{all}{Plot all types of SNVs, default only C>X and T>X are plotted.}

\item{alpha}{adjust the alpha level of the points}

\item{legend}{If TRUE display a figure legend}

\item{colours}{a function specifying the colours to use in the plot, see the
default value (\code{kataegis.colours}) for an example.}

\item{...}{passed to \code{plot} function}
}
\value{
invisibly return \code{data.frame} of SNVs plotted.
}
\description{
Create a kataegis "rainfall" plot.
}
\details{
In its simplest form (\code{plot(x)}), this function produces a rainfall
plot on the current graphical device.  By default, the whole genome is
displayed.  This can be changed by providing either the \code{chr} parameter
(to show a specific chromosome) or the \code{chr}, \code{start} and
\code{end} parameters to show a specific region of a chromosome (note that
all three parameters must be provided).
}
\examples{
\dontrun{
vcf.file <- system.file('extdata', 'example.vcf', package = 'GenomicVis')
x <- kategis(vcf.file)
main <- sprintf("Kataegis: \%d", length(x$kataegis))
plot(x, main = main)
}
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}
\keyword{methods}

