% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotKataegis.R
\name{plotKataegis}
\alias{plotKataegis}
\title{Plot a kataegis rainfall figure.}
\usage{
plotKataegis(x, chr = NULL, all = FALSE, alpha.f = 1,
  colour.map = kataegis.colours, main = NULL, newpage = TRUE,
  panel_colour = "grey90", point_size = unit(4, "point"))
}
\arguments{
\item{x}{A \code{GRanges} instance containing SNVs.}

\item{chr}{Create a rainfall plot for a single chromosome specified by this
parameter.}

\item{all}{all}

\item{alpha.f}{number in (0, 1] specifiying the alpha level with which to
draw the points.}

\item{colour.map}{A list specifying the colours to use in the figure.  See
\code{kataegis.colours} for exact format.}

\item{main}{Title for plot}

\item{newpage}{logical. If TRUE create figure on a new device, otherwise
print to current device.}

\item{panel_colour}{Colour of background panels indicating chromosomes.}

\item{point_size}{specify the size of individual points.}
}
\value{
None
}
\description{
Create a rainfall plot showing potential regions of kataegis.  The input
\code{GRanges} instance should only contain SNV (i.e., no INDELs) that have
a single ALT allele.  The input \code{GRanges} instance must also contain a
valid \code{Seqinfo} class that gives the lengths of the chromosomes.  The
easiest way to obtain this from a VCF file is via the \code{read.vcf}
function from this package.
}
\examples{
library(VariantAnnotation)
vcf.file <- system.file('extdata', 'example.vcf', package = 'GenomicVis')
vcf <- read.vcf(vcf.file, 'GRCh37')
x <- rowData(vcf)
plotKataegis(x)
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}

