% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prepare.data}
\alias{prepare.data}
\title{Prepare data for plotting.}
\usage{
prepare.data(chr, snps, segments, svs)
}
\arguments{
\item{chr}{chr}

\item{snps}{snps}

\item{segments}{segments}

\item{svs}{svs}
}
\description{
This is an internal helper function for \code{cnv.plot}.
}
\details{
The three main \code{data.frame}s are subset to only include the reqested
chromosome and a \code{GRange} object is created from the segmentation
data.
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}

