% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.kataegisList}
\alias{print.kataegisList}
\title{Print function for kataegisList}
\usage{
\method{print}{kataegisList}(x, ...)
}
\arguments{
\item{x}{a \code{kataegisList} produced by \code{kataegis}}

\item{...}{passed to what?}
}
\value{
none
}
\description{
Print function for kataegisList
}
\details{
What is a kataegis plot and how is kataegis defined?
}
\examples{
\dontrun{
vcf.file <- system.file('extdata', 'example.vcf',
  package = 'GenomicVis')
x <- kataegis(vcf.file)
print(x)
}
}
\author{
Jonatha Ellis <jonathan.j.ellis@gmail.com>
}
\keyword{methods}

