% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.breakdancer.R
\name{read.breakdancer}
\alias{read.breakdancer}
\title{Illustrative function to read the output of Breakdancer.}
\usage{
read.breakdancer(f, chrom = NA, normal.regex = "NORMAL",
  tumour.regex = "TUMOUR")
}
\arguments{
\item{f}{The Breakdancer output file name}

\item{chrom}{The chromosome to process}

\item{normal.regex}{A regular expression used on the input file names to
identify the normal sample.}

\item{tumour.regex}{A regular expression used on the input file names to
identify the tumour sample.}
}
\value{
A \code{data.frame} suitable as input to \code{cnv.plot}
}
\description{
This function reads a Breakdancer output file and returns a
\code{data.frame} suitable as input to \code{cnv.plot}.  This only serves
as an example, Breakdancer output is not required to run \code{cnv.plot},
the structural variant data may come from any source.
}
\examples{
\dontrun{
read.breakdancer()
}
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}

