% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.gap}
\alias{read.gap}
\title{Illustrative function to read GAP output.}
\usage{
read.gap(file, annot, sample.id = NULL)
}
\arguments{
\item{file}{GAP output file}

\item{annot}{GAP output annotation file}

\item{sample.id}{The name of the sample to process}
}
\value{
A \code{data.frame} instance that can be used as input for
\code{cnv.plot}.
}
\description{
This function is provided as an example of reading output from GAP and
creating the required \code{data.frame} for \code{cnv.plot}.
}
\examples{
\dontrun{
read.gap()
}
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}

