% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.illumina}
\alias{read.illumina}
\title{Illustrative function to an Illumina GenomeStudio file.}
\usage{
read.illumina(file, snp.name = "SNP Name", sample.name = "Sample ID",
  chr.name = "Chr", pos.name = "Position", baf.name = "B Allele Freq",
  lrr.name = "Log R Ratio")
}
\arguments{
\item{file}{The GenomeStudio file to process.}

\item{snp.name}{The name of the SNP ID column.}

\item{sample.name}{The name of the sample ID column.}

\item{chr.name}{The name of the chromosome column.}

\item{pos.name}{The name of the SNP position column.}

\item{baf.name}{The name of the B allele frequency column.}

\item{lrr.name}{The name of the log R ratio column.}
}
\value{
A \code{data.table} instance with six columns: \code{SNP.Name},
\code{Sample.ID}, \code{Chr}, \code{Position}, \code{BAF} and \code{LRR}.
}
\description{
This function is an example of reading the output of GenomeStudio and
producing a \code{data.frame} required by \code{cnv.plot}.
}
\details{
The input file is the output (i.e., final report) from GenomeStudio.  It
must contain at least six columns with the SNP name, sample ID, chromosome,
position B allele frequency and log R ratio in it.  The exact column name
must match the parameters \code{chr.name}, \code{pos.name}, \code{baf.name}
and \code{lrr.name} respectively.
}
\examples{
\dontrun{
x <- read.illumina('LC1_TUMOUR_A_FinalReport.txt')
}
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}

