% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.snpeff.vcf}
\alias{read.snpeff.vcf}
\title{Read a SnpEff annotated VCF file.}
\usage{
read.snpeff.vcf(f, genome, tumour.id, exclude.filtered = TRUE)
}
\arguments{
\item{f}{f}

\item{genome}{genome}

\item{tumour.id}{tumour id}

\item{exclude.filtered}{logical, exclude filtered variants (defined as
variants with some thing other that 'PASS' or '.' in their FILTER
column).}
}
\value{
A \code{data.frame}.
}
\description{
Illustrative function to read gene names from a VCF file that has been
annotated with SnpEff.
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}

