% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.snpeff.vcfs}
\alias{read.snpeff.vcfs}
\title{Illustrative function to read multiple VCFs}
\usage{
read.snpeff.vcfs(vcf.files, genome, sample.names = NULL)
}
\arguments{
\item{vcf.files}{vcf.files}

\item{genome}{genome}

\item{sample.names}{sample.names}
}
\value{
A \code{data.frame} suitable as input to \code{snv.heatmap}.
}
\description{
Illustrative function to read multiple VCFs and create an appropriate
input \code{data.frame} for \code{snv.heatmap}.
}

