% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenomicVis-package.R
\docType{data}
\name{SNPExample}
\alias{SNPExample}
\title{An example of a SNP data frame used by \code{cnv.plot}.}
\description{
An example of a SNP data frame used by \code{cnv.plot}.
}
\keyword{datasets}

