% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snv.clustering.R
\name{snv.clustering}
\alias{snv.clustering}
\title{Cluster samples based on SNVs.}
\usage{
snv.clustering(vcf.files, genome, sample.names = NULL, ...)
}
\arguments{
\item{vcf.files}{A character vector containing the paths to the VCF files to
process.}

\item{genome}{A character string identifying the genome}

\item{sample.names}{A character vector of samples names corresponding to
each VCF file, by default the file names of the VCFs will be used as sample
names.}

\item{...}{Additional parameters pass to \code{read.vcf}.  These can be used
to apply additional filtering of the SNVs, see the documentation of
\code{read.vcf} for further details.}
}
\value{
Invisible returns an object of class \code{hclust} that describes
the tree produced by the clustering process.
}
\description{
Perform hierarchical clustering of samples based on their somatic SNVs.
}
\details{
Each vcf file should contain called somatic SNVs.  use exclude.filtered =
TRUE (the default) to only use SNVs with 'PASS' or '.' in the FILTER column.
}
\examples{
vcf.files <- c(
  system.file('extdata', 'LC1_TUMOUR_A.vcf', package = 'GenomicVis'),
  system.file('extdata', 'LC1_TUMOUR_B.vcf', package = 'GenomicVis'),
  system.file('extdata', 'LC2_TUMOUR_A.vcf', package = 'GenomicVis'),
  system.file('extdata', 'LC2_TUMOUR_B.vcf', package = 'GenomicVis'),
  system.file('extdata', 'LC3_TUMOUR_A.vcf', package = 'GenomicVis'),
  system.file('extdata', 'LC3_TUMOUR_B.vcf', package = 'GenomicVis'))
sample.names <- tools::file_path_sans_ext(basename(vcf.files))
snv.clustering(vcf.files, sample.names, genome = 'hg19')
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}

