% Generated by roxygen2 (4.0.1): do not edit by hand
\name{snv.heatmap}
\alias{snv.heatmap}
\title{Plot a SNV heatmap.}
\usage{
snv.heatmap(x, width = 1000, height = 5000, margins = c(5, 12), ...)
}
\arguments{
\item{x}{A \code{data.frame} with (at least) two columns name \code{Gene}
and \code{Tumour}.}

\item{width}{width}

\item{height}{height}

\item{margins}{margins}

\item{...}{?}
}
\description{
description
}
\details{
details
}
\examples{
set.seed(1000)
genes <- paste0('Gene_', toupper(letters)[1:20])
tumours <- paste0('LC1_', toupper(letters)[1:4])
x <- sapply(1:length(genes),
  function(i) sample(tumours, sample(length(tumours))[1]))
names(x) <- genes
y <- as.data.frame(unlist(x))
dat <- data.frame(Gene = substr(rownames(y), 1, 6), Tumour = y[, 1])
snv.heatmap(dat, margins = c(5, 8))
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}

