% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarise.mutations.R
\name{summarise.mutations}
\alias{summarise.mutations}
\title{Summarise the types of mutations in a region of kataegis.}
\usage{
summarise.mutations(x, pos = 1)
}
\arguments{
\item{x}{a \code{kataegisList} object, usually created by \code{kataegis}.}

\item{pos}{the index of the kataegis region in \code{x} to summarise.}
}
\value{
returns a contingency table summarising the types of mutation found
in the specified region of kataegis.
}
\description{
Summarise the types of mutations in a region of kataegis.
}
\examples{
\dontrun{
x <- kataegis(system.file('extdata', 'example.vcf', package = 'GenomicVis')
}
data(KataegisExample)
summarise.mutations(KataegisExample, pos = 2)
}
\author{
Jonathan Ellis <jonathan.j.ellis@gmail.com>
}

