% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vcf.venn}
\alias{vcf.venn}
\title{Create a venn diagram of SNV overlap in a group of VCF files.}
\usage{
vcf.venn(vcf.files, genome, sample.names = NULL, ...)
}
\arguments{
\item{vcf.files}{A character vector of VCF file to process.}

\item{sample.names}{A character vector of samples names for each of the VCF
files.  Must be the same length as \code{vcf.files}.  By default the VCF
files names are used as sample names.}

\item{genome}{A character string identifying the genome}

\item{...}{Additionally parameters to pass to \code{read.vcf}.}
}
\value{
A list containing:
\item{venn}{An object of class \code{Venn}, see documentation of package
\code{Vennerable} for further details and methods.}
\item{data}{The data list that was used to create the \code{Venn} instance.}
}
\description{
There is a maximum number of samples (nine) that can be included in a venn
diagram (this limit is imposed by the R package producing the figures).
}
\examples{
\dontrun{
f1 <- system.file('extdata', 'LC1_TUMOUR_A.vcf', package = 'GenomicVis')
f2 <- system.file('extdata', 'LC1_TUMOUR_B.vcf', package = 'GenomicVis')
vcf.files <- c(f1, f2)
sample.names <- c('LC1_A', 'LC1_B')
v <- vcf.venn(vcf.files, 'GRCh37', sample.names)
plot(v$venn)
}
}

