% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cnv.merge}
\alias{cnv.merge}
\title{Merge CNVs}
\usage{
cnv.merge(cnv, samples)
}
\arguments{
\item{cnv}{cnv}

\item{samples}{samples}
}
\value{
?
}
\description{
This is an internal helper function for \code{cnv.heatmap}.
}
\author{
Lutz Krause <lutz.krause@qimrberghofer.edu.au>
}

