% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{hg19Genes}
\alias{hg19Genes}
\title{A \code{GRanges} instance that contains hg19 gene locations taken from
RefSeq with corresponding gene symbols in the \code{mcol} attribute.}
\description{
A \code{GRanges} instance that contains hg19 gene locations taken from
RefSeq with corresponding gene symbols in the \code{mcol} attribute.
}
\keyword{datasets}

