% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{KataegisExample}
\alias{KataegisExample}
\title{An example of output from the \code{kataegis} function.}
\format{A data frame with x rows and y variables}
\description{
This dataset serves as an example of the output of the \code{kataegis}
function, and it can be used for creating kataegis rainfall plots.
}
\keyword{datasets}

